move app quit to window close event instead
